My warning vigilantism for the day: perform some explicit casts to remove the implicit (yet likely correct) casting going on.
* subversion/libsvn_fs_fs/caching.c
  (manifest_deserialize): Cast a few values to int where required by the
    APR api.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1040115 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs_fs/caching.c
@@ -134,10 +134,10 @@ manifest_deserialize(void **out,
                      apr_pool_t *pool)
 {
   apr_array_header_t *manifest = apr_array_make(pool,
-                                                data_len / sizeof(apr_off_t),
-                                                sizeof(apr_off_t));
+                                       (int) (data_len / sizeof(apr_off_t)),
+                                       (int) sizeof(apr_off_t));
   memcpy(manifest->elts, data, data_len);
-  manifest->nelts = data_len / sizeof(apr_off_t);
+  manifest->nelts = (int) (data_len / sizeof(apr_off_t));
   *out = manifest;
 
   return SVN_NO_ERROR;
